Move some cross_canister_call specific stuff
