+++ src/compiler/typescript_to_rust/azle_generate/src/body/ic_object/functions/cross_canister_calls/call.rs
@@ -2,8 +2,6 @@ use cdk_framework::act::node::candid::{service::Method, Service};
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 
-use crate::generators::ic_object;
-
 pub fn generate(
     canister: &Service,
     method: &Method,
@@ -14,8 +12,8 @@ pub fn generate(
     let call_function_name_string = format!("_azle_call_{}_{}", canister.name, method.name);
     let call_function_name_ident = format_ident!("{}", call_function_name_string);
     let call_wrapper_fn_name = format_ident!("{}_wrapper", call_function_name_string);
-    let param_variables = ic_object::generate_param_variables(method, &canister.name);
-    let args = ic_object::generate_args_list(method);
+    let param_variables = super::generate_param_variables(method, &canister.name);
+    let args = super::generate_args_list(method);
 
     quote! {
         fn #call_wrapper_fn_name(
