+++ src/compiler/typescript_to_rust/azle_generate/src/body/ic_object/functions/cross_canister_calls/notify_with_payment128.rs
