+++ src/compiler/typescript_to_rust/azle_generate/src/lib.rs
@@ -12,7 +12,6 @@ mod errors;
 mod header;
 mod ts_ast;
 mod ts_keywords;
-mod utils;
 mod vm_value_conversion;
 
 pub fn generate_canister(ts_file_names: &Vec<&str>, main_js: String) -> TokenStream {
