--- src/compiler/typescript_to_rust/azle_generate/src/utils.rs
@@ -1,7 +1,0 @@
-pub trait ToU8 {
-    fn to_u8(&self) -> Result<u8, String>;
-}
-
-pub trait ToU32 {
-    fn to_u32(&self) -> Result<u32, String>;
-}
